Summary 250320
Tutorial Function Added
Tutorial Stage Added

Game UI convenience improved.

+++ Server/Data/SinglePlayScenario/Stage1001.cs
@@ -5,7 +5,7 @@ namespace Server.Data.SinglePlayScenario;
 
 public class Stage1001 : Stage
 {
-    private readonly Dictionary<float, MonsterStatue> _statues = new();
+    private readonly Dictionary<int, MonsterStatue> _statues = new();
     
     public override void Spawn(int round)
     {
